requirements.txt
oss2>=2.17.0
Pillow>=9.0.0